requirements.txt
Deprecated==1.2.14
img2pdf==0.5.1
lxml==5.3.0
packaging==24.2
pikepdf==9.4.0
pillow==11.0.0
PyPDF2==3.0.1
wrapt==1.16.0
